using LengthPrefixCoder instead of adding a new one
